Move ConnectionsInfo into header & source files
